decouple accounting into another plugin
