+++ utilities.go
@@ -5,22 +5,7 @@ import (
 	"fmt"
 	"hash/fnv"
 	"strings"
-
-	"golang.org/x/crypto/bcrypt"
 )
-
-// Hashes password using Bcrypt algorithm
-func hashString(password string) (string, error) {
-	bytes, err := bcrypt.GenerateFromPassword([]byte(password), 14)
-
-	return string(bytes), err
-}
-
-// Compares hash and password.
-func compareHash(password, hash string) bool {
-	err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password))
-	return err == nil
-}
 
 // Finds `T` value of []T.
 func Find[T comparable](collection []T, value T) bool {
@@ -176,37 +161,6 @@ func FilterByType[T any, TCollection any](collection []TCollection) []T {
 		}
 	}
 	return result
-}
-
-func saveRoomsRecursively(engine *Akevitt, room *Room, visited []string) error {
-	if visited == nil {
-		visited = make([]string, 0)
-	}
-
-	if room == nil {
-		return errors.New("room is nil")
-	}
-
-	fmt.Printf("Loading Room: %s\n", room.Name)
-
-	engine.rooms[room.GetKey()] = room
-
-	visited = append(visited, room.Name)
-
-	for _, v := range room.Exits {
-		r := v.Room
-
-		if Find[string](visited, r.Name) {
-			continue
-		}
-
-		err := saveRoomsRecursively(engine, r, visited)
-
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 func hash(s string) uint64 {
